backend/requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2
bs4
selenium
webdriver_manager
lxml
pandas
pathlib
